+++ modules/features/podcasts/src/main/java/au/com/shiftyjelly/pocketcasts/podcasts/view/podcast/PodcastSettingsPage.kt
@@ -10,8 +10,12 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
@@ -20,20 +24,19 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import au.com.shiftyjelly.pocketcasts.compose.bars.NavigationButton
 import au.com.shiftyjelly.pocketcasts.compose.bars.ThemedTopAppBar
+import au.com.shiftyjelly.pocketcasts.compose.components.FormFieldDialog
 import au.com.shiftyjelly.pocketcasts.compose.navigation.navigateOnce
 import au.com.shiftyjelly.pocketcasts.compose.navigation.slideInToEnd
 import au.com.shiftyjelly.pocketcasts.compose.navigation.slideInToStart
 import au.com.shiftyjelly.pocketcasts.compose.navigation.slideOutToEnd
 import au.com.shiftyjelly.pocketcasts.compose.navigation.slideOutToStart
 import au.com.shiftyjelly.pocketcasts.compose.theme
-import au.com.shiftyjelly.pocketcasts.models.to.AutoArchiveAfterPlaying
-import au.com.shiftyjelly.pocketcasts.models.to.AutoArchiveInactive
-import au.com.shiftyjelly.pocketcasts.models.to.AutoArchiveLimit
 import au.com.shiftyjelly.pocketcasts.podcasts.viewmodel.PodcastSettingsViewModel
-import au.com.shiftyjelly.pocketcasts.repositories.playlist.PlaylistPreview
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.Feature
 import au.com.shiftyjelly.pocketcasts.utils.featureflag.FeatureFlag
 import au.com.shiftyjelly.pocketcasts.views.helper.ToolbarColors
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.flow.StateFlow
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
@@ -61,8 +64,10 @@ internal fun PodcastSettingsPage(
     onChangeVolumeBoost: (Boolean) -> Unit,
     onDecrementSkipFirst: () -> Unit,
     onIncrementSkipFirst: () -> Unit,
+    onChangeSkipFirst: (String) -> Unit,
     onDecrementSkipLast: () -> Unit,
     onIncrementSkipLast: () -> Unit,
+    onChangeSkipLast: (String) -> Unit,
     onAddPodcastToPlaylists: (List<String>) -> Unit,
     onRemovePodcastFromPlaylists: (List<String>) -> Unit,
     onUnfollow: () -> Unit,
@@ -70,6 +75,9 @@ internal fun PodcastSettingsPage(
     modifier: Modifier = Modifier,
     navController: NavHostController = rememberNavController(),
 ) {
+    var isSkipFirstDialogOpen by remember { mutableStateOf(false) }
+    var isSkipLastDialogOpen by remember { mutableStateOf(false) }
+
     val backStackEntry by navController.currentBackStackEntryAsState()
     val toolbarTitle = when (backStackEntry?.destination?.route) {
         PodcastSettingsRoutes.HOME, null -> podcastTitle
@@ -146,6 +154,16 @@ internal fun PodcastSettingsPage(
                     onChangePlaylistSettings = {
                         navController.navigateOnce(PodcastSettingsRoutes.PLAYLISTS)
                     },
+                    onClickSkipFirst = {
+                        if (!isSkipLastDialogOpen) {
+                            isSkipFirstDialogOpen = true
+                        }
+                    },
+                    onClickSkipLast = {
+                        if (!isSkipFirstDialogOpen) {
+                            isSkipLastDialogOpen = true
+                        }
+                    },
                     onUnfollow = onUnfollow,
                 )
             }
@@ -193,6 +211,48 @@ internal fun PodcastSettingsPage(
             }
         }
     }
+
+    if (isSkipFirstDialogOpen) {
+        ChangeSkipDurationDialog(
+            title = stringResource(LR.string.podcast_settings_skip_first),
+            initialDuration = uiState?.podcast?.startFromSecs?.seconds ?: Duration.ZERO,
+            onConfirm = { value ->
+                onChangeSkipFirst(value)
+                isSkipFirstDialogOpen = false
+            },
+            onDismiss = { isSkipFirstDialogOpen = false },
+        )
+    }
+
+    if (isSkipLastDialogOpen) {
+        ChangeSkipDurationDialog(
+            title = stringResource(LR.string.podcast_settings_skip_last),
+            initialDuration = uiState?.podcast?.skipLastSecs?.seconds ?: Duration.ZERO,
+            onConfirm = { value ->
+                onChangeSkipLast(value)
+                isSkipLastDialogOpen = false
+            },
+            onDismiss = { isSkipLastDialogOpen = false },
+        )
+    }
+}
+
+@Composable
+private fun ChangeSkipDurationDialog(
+    title: String,
+    initialDuration: Duration,
+    onConfirm: (String) -> Unit,
+    onDismiss: () -> Unit,
+) {
+    FormFieldDialog(
+        title = title,
+        placeholder = stringResource(LR.string.seconds_label),
+        initialValue = initialDuration.inWholeSeconds.toString(),
+        keyboardType = KeyboardType.Number,
+        onConfirm = onConfirm,
+        onDismissRequest = onDismiss,
+        isSaveEnabled = { value -> value.toIntOrNull()?.takeIf { it >= 0 } != null },
+    )
 }
 
 private object PodcastSettingsRoutes {