+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/StorageSettingsPage.kt
@@ -11,11 +11,7 @@ import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.selection.toggleable
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.MaterialTheme
-import androidx.compose.material.OutlinedTextField
-import androidx.compose.material.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -24,13 +20,10 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.Dp
@@ -38,14 +31,14 @@ import androidx.compose.ui.unit.dp
 import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.compose.CallOnce
 import au.com.shiftyjelly.pocketcasts.compose.bars.ThemedTopAppBar
-import au.com.shiftyjelly.pocketcasts.compose.components.DialogButtonState
-import au.com.shiftyjelly.pocketcasts.compose.components.DialogFrame
+import au.com.shiftyjelly.pocketcasts.compose.components.DialogButtonProperties
+import au.com.shiftyjelly.pocketcasts.compose.components.FormFieldDialog
 import au.com.shiftyjelly.pocketcasts.compose.components.ProgressDialog
 import au.com.shiftyjelly.pocketcasts.compose.components.SettingRadioDialogRow
 import au.com.shiftyjelly.pocketcasts.compose.components.SettingRow
 import au.com.shiftyjelly.pocketcasts.compose.components.SettingRowToggle
 import au.com.shiftyjelly.pocketcasts.compose.components.SettingSection
-import au.com.shiftyjelly.pocketcasts.compose.components.TextP40
+import au.com.shiftyjelly.pocketcasts.compose.components.SimpleDialog
 import au.com.shiftyjelly.pocketcasts.compose.components.TextP50
 import au.com.shiftyjelly.pocketcasts.compose.preview.ThemePreviewParameterProvider
 import au.com.shiftyjelly.pocketcasts.compose.theme
@@ -55,8 +48,6 @@ import au.com.shiftyjelly.pocketcasts.settings.viewmodel.StorageSettingsViewMode
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.utils.Util
 import au.com.shiftyjelly.pocketcasts.utils.log.LogBuffer
-import java.util.*
-import kotlinx.coroutines.delay
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 @Composable
@@ -82,7 +73,6 @@ fun StorageSettingsPage(
     if (showProgressDialog) {
         ProgressDialog(
             text = stringResource(LR.string.settings_storage_move_podcasts),
-            onDismiss = { showProgressDialog = false },
         )
     }
 
@@ -266,59 +256,14 @@ private fun StorageFolderRow(
                 .clickable { showDialog = true },
         ) {
             if (showDialog) {
-                val focusRequester = remember { FocusRequester() }
-                LaunchedEffect(Unit) {
-                    // delay apparently needed to ensure the soft keyboard opens
-                    delay(100)
-                    focusRequester.requestFocus()
-                }
-
-                var value by remember {
-                    mutableStateOf(
-                        TextFieldValue(
-                            text = storageFolderState.summary ?: "",
-                        ),
-                    )
-                }
-
-                val onFinish = {
-                    storageFolderState.onStateChange(value.text)
-                    showDialog = false
-                }
-
-                DialogFrame(
+                FormFieldDialog(
                     title = stringResource(LR.string.settings_storage_custom_folder_location),
-                    buttons = listOf(
-                        DialogButtonState(
-                            text = stringResource(LR.string.cancel).uppercase(
-                                Locale.getDefault(),
-                            ),
-                            onClick = { showDialog = false },
-                        ),
-                        DialogButtonState(
-                            text = stringResource(LR.string.ok),
-                            onClick = onFinish,
-                        ),
-                    ),
+                    placeholder = stringResource(LR.string.settings_storage_custom_folder_location),
+                    initialValue = storageFolderState.summary.orEmpty(),
+                    keyboardType = KeyboardType.Text,
+                    onConfirm = { value -> storageFolderState.onStateChange(value) },
                     onDismissRequest = { showDialog = false },
-                ) {
-                    OutlinedTextField(
-                        value = value,
-                        onValueChange = {
-                            value = it
-                        },
-                        colors = TextFieldDefaults.textFieldColors(
-                            textColor = MaterialTheme.theme.colors.primaryText01,
-                            placeholderColor = MaterialTheme.theme.colors.primaryText02,
-                            backgroundColor = MaterialTheme.theme.colors.primaryUi01,
-                        ),
-                        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Text),
-                        keyboardActions = KeyboardActions { onFinish() },
-                        modifier = Modifier
-                            .padding(horizontal = 24.dp)
-                            .focusRequester(focusRequester),
-                    )
-                }
+                )
             }
         }
     }
@@ -368,27 +313,18 @@ private fun AlertDialogView(
     alertDialogState: StorageSettingsViewModel.AlertDialogState,
     onDismiss: () -> Unit,
 ) {
-    DialogFrame(
+    SimpleDialog(
         title = alertDialogState.title,
-        buttons = alertDialogState.buttons.map {
-            DialogButtonState(
+        body = alertDialogState.message.orEmpty(),
+        onDismissRequest = onDismiss,
+        buttonProperties = alertDialogState.buttons.map {
+            DialogButtonProperties(
                 text = it.text,
                 onClick = {
                     it.onClick()
                     onDismiss()
                 },
             )
-        },
-        onDismissRequest = { onDismiss() },
-        content = {
-            alertDialogState.message?.let {
-                TextP40(
-                    text = it,
-                    modifier = Modifier
-                        .padding(bottom = 12.dp)
-                        .padding(horizontal = 24.dp),
-                )
-            }
         },
     )
 }