+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 7.102
 -----
 *   Updates
+    *   Bring back podcast setting skip first/last dialog
+        ([#4655](https://github.com/Automattic/pocket-casts-android/pull/4655))
     *   Exclude empty searches from history
         ([#4660](https://github.com/Automattic/pocket-casts-android/pull/4660))